+++ app/generator/content/part2/get-table.js
@@ -1,40 +1,3 @@
-const part2 = (statement) => {
-  const reductions = mapReductions(statement.funding)
-
-  return [
-    { text: 'Part 2. Calculation', style: 'header2' },
-    { text: `We calculated the total ${statement.scheme.shortName} payment amount by adding together payments for all the standards which are part of your agreement, as shown in this table.` },
-    getTable(statement.funding, reductions),
-    {
-      text: [
-        'You can check which land parcels are part of each standard in the Rural Payments service at ',
-        { text: 'www.gov.uk/claim-rural-payments', link: 'https://www.gov.uk/claim-rural-payments', style: 'link' },
-        '.'
-      ]
-    },
-    ...getReductions(reductions)
-  ]
-}
-
-const mapReductions = (funding) => {
-  const reductions = []
-  let i = 1
-  funding.forEach(x => {
-    if (x.reductions) {
-      x.reductions.forEach(y => {
-        if (!reductions.find(z => z.reason === y.reason)) {
-          reductions.push({
-            id: i,
-            reason: y.reason
-          })
-          i++
-        }
-      })
-    }
-  })
-  return reductions
-}
-
 const getTable = (funding, reductions) => {
   return {
     layout: {
@@ -99,22 +62,4 @@ const mapReductionIds = (fundReductions, reductions) => {
   })).join('')}`
 }
 
-const getReductions = (reductions) => {
-  if (!reductions.length) {
-    return ['']
-  }
-  return [
-    { text: 'Reason for reductions', style: 'header3' },
-    ...getReductionList(reductions)
-  ]
-}
-
-const getReductionList = (reductions) => {
-  const list = []
-  reductions.forEach(x => {
-    list.push(`(${x.id}) ${x.reason}`)
-  })
-  return list
-}
-
-module.exports = part2
+module.exports = getTable
